feat: add logic for buttons, animations and fetching data from db

+++ feature/src/main/java/com/example/feature/presentation/home/presentation/screen/HomeScreen.kt
@@ -40,7 +40,7 @@ import okio.IOException
 @OptIn(ExperimentalMaterialApi::class)
 @Composable
 fun HomeScreen (
-    onPhotoClick: () -> Unit,
+    onPhotoClick: (Long) -> Unit,
     viewModel: HomeViewModel = hiltViewModel(),
 ) {
     Column(
